refactor: update state variable types in DeployBase and FarcasterCrowdfund contracts for better efficiency; adjust donation handling and add new test cases for edge scenarios

+++ script/DeployBase.s.sol
@@ -12,7 +12,7 @@ contract DeployBase is Script {
     string public constant BASE_URI = "https://crowdfund.seedclub.com/nfts/";
     
     // Maximum crowdfund duration (7 days)
-    uint256 public constant MAX_DURATION = 7 days;
+    uint64 public constant MAX_DURATION = 7 days;
 
     function run() public {
         // Retrieve deployer private key from env variable
